move branch lessons rendering to a single component:lessonContainer.jsx and remove logic from branches files
